+++ graalpython/com.oracle.graal.python/src/com/oracle/graal/python/compiler/OpCodes.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2021, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -882,8 +882,8 @@ public enum OpCodes {
         this.producesStackItems = quickens.producesStackItems;
         this.generalizesTo = generalizesTo;
         this.quickens = quickens;
-        quickens.quickenInputTypes |= inputType;
-        quickens.quickenOutputTypes |= outputType;
+        quickens.quickenInputTypes |= (byte) inputType;
+        quickens.quickenOutputTypes |= (byte) outputType;
     }
 
     public byte canQuickenInputTypes() {
